Write Java Codes
1) Write a java code with the class named ‘acad’ and a method ‘main’. Hard Code the program
with two integers and print the sum of those two.
2) Rewrite the above code, where, inputs are provided by the user at runtime and the output is
printed.
3) Write a program with method name sum() that accepts two parameters from user and print
the sum of two numbers. Output format should be as:
First number is:
Second number is:
Sum is:
4) Write a program to accepts two numbers from stdin and find all the odd as well as even
numbers present in between them.
5) Joe is scared to go to school. When her dad asked the reason, joe said she is unable to
complete the task given by her teacher. The task was to find the “first 10 multiples” of the
number entered from stdin . Eg:
Input: 3
O/p:
3 x 1 = 3
3 x 2 = 6
………
………
…….
….
3 x 10 = 30
Help Joe in completing the task!
6) Write a program consisting method sum() and demonstrate the concept of method
overloading using this method.
7) Can you overload a method with same return type.? Explain your answer with proper logic.
8) Write a program in java using Arrays, that sorts the element in descending order.